requirements.txt
# Lead Scoring Agent - Dependencies
# Python 3.9+

streamlit>=1.28.0
pandas>=2.0.0
python-dotenv>=1.0.0
